Short story
I wrote a quick short story (~600 words), to explore the worldbuilding, extend the lore, and generally play around with the setting. The writing is fairly terse, as it more closely approximates game text. You probably wouldn't give a player this sort of dump, but the conversations might play out like this.

It also has a few ideas for gameplay ideas.